backend/requirements.txt
asgiref==3.7.2
agora_token_builder==1.0.0
autobahn==24.4.2
certifi==2025.1.31
cffi==1.17.1
channels==4.1.0
channels-redis==4.1.0
charset-normalizer==3.4.1
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
razorpay==1.4.2
redis==5.0.1
requests==2.32.3
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
urllib3==2.4.0
websockets==15.0.1
whitenoise==6.6.0
